Implement priority support
Allow setting the priority of a service using one of two syntaxes:

service "foo" do
    priority 20
end

service "foo" do
    priority { 2 => [ :start, 20 ], 3 => [ :stop, 80 ] }
end

The former will enable the service with priority 20 in the default
runlevel.  It will be ignored if the system does not have a priority
or runlevel concept.  The second syntax will set the start priority to
20 in runlevel 2 and stop priority to 80 in runlevel 3.